requirements.txt
pyTelegramBotAPI==4.23.0
googletrans==3.1.0a0
easyocr==1.7.1
opencv-python-headless==4.10.0.84
Pillow==10.4.0
python-dotenv==1.0.0
numpy==1.26.4
google-cloud-vision==3.4.4